django_trans/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2-binary==2.8.6
gunicorn==20.1.0
django-cors-headers==3.7.0
channels==3.0.5
django-sslserver==0.22
channels_redis==4.2.0